condense serializers
