+++ modules/vye/app/serializers/vye/vye_serializer.rb
@@ -16,4 +16,35 @@ module Vye
       @resource&.status
     end
   end
+
+  class ClaimantLookupSerializer < VyeSerializer
+    def serializable_hash
+      {
+        claimant_id: @resource&.claimant_id
+      }
+    end
+  end
+
+  class ClaimantVerificationSerializer < VyeSerializer
+    def serializable_hash
+      {
+        claimant_id: @resource&.claimant_id,
+        delimiting_date: @resource&.delimiting_date,
+        enrollment_verifications: @resource&.enrollment_verifications,
+        verified_details: @resource&.verified_details,
+        payment_on_hold: @resource&.payment_on_hold
+      }
+    end
+  end
+
+  class VerifyClaimantSerializer < VyeSerializer
+    def serializable_hash
+      {
+        claimant_id: @resource&.claimant_id,
+        delimiting_date: @resource&.delimiting_date,
+        verified_details: @resource&.verified_details,
+        payment_on_hold: @resource&.payment_on_hold
+      }
+    end
+  end
 end
